Change Header for Alpha.Data.gov
Hey, guys.  This is the feedback I suggested via email a while back. Now incorporated in this pull request.

"A collection of open data from the government, private sector, and non-profits that are fueling anew economy."

becomes:

"A collection of open data from governments, companies, and non-profits is fueling anew economy."